getl.utils.Config: the configuration management class has been reworked, the implementation of processing configuration files has been rendered in the сonfig.ConfigFiles class, in the future it is planned to add modules for storing the configuration in databases or working with other solutions.

+++ src/main/groovy/getl/files/HDFSManager.groovy
@@ -55,7 +55,7 @@ class HDFSManager extends Manager {
     /**
      * Server port
      */
-    public Integer getPort () { (params.port != null)?params.port:8022 }
+    public Integer getPort () { (params.port != null)?(params.port as Integer):8022 }
     public void setPort (Integer value) { params.port = value }
 
     /**
